backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0

# Database - PostgreSQL
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Database - Neo4j
neo4j==5.20.0

# Cache - Redis
redis==5.0.4
hiredis==2.3.2

# Task Queue
celery==5.4.0
flower==2.0.1

# Parsing & AST
tree-sitter==0.22.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0
tree-sitter-java==0.21.0
tree-sitter-go==0.21.0
tree-sitter-rust==0.21.1
gitpython==3.1.43

# Embeddings & ML
openai==1.35.0
numpy==1.26.4
faiss-cpu==1.8.0

# NLP & Text
spacy==3.7.4
nltk==3.8.1

# Validation & Config
pydantic==2.7.3
pydantic-settings==2.3.1
python-dotenv==1.0.1

# Observability
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
prometheus-client==0.20.0
structlog==24.2.0

# Utilities
tenacity==8.3.0
xxhash==3.4.1
pyyaml==6.0.1
aiofiles==23.2.1
